refactor(docs-retrieval): extract cover-page from content-fetcher
Move the learning-path cover-page renderer out of content-fetcher.ts into
a focused sibling module `content-fetcher/cover-page.ts` (PR 4 of the
content-fetcher modularization stack).

## What moved (verbatim, no logic change)

- `simpleMarkdownToHtml` + `inlineMarkdown` + `escapeHtmlEntities` — the
  minimal markdown→HTML renderer for cover content (headings, paragraphs,
  unordered lists, sanitized inline links)
- `LEARNING_PATH_ICON_SVG`, `wrapInOrangeOutlineList` — the "what to expect"
  card chrome
- `injectJourneyExtrasIntoJsonGuide` + `wrapExpectBlockInOrangeOutline` +
  `splitAtNextHeading` + the `EXPECT_HEADING_RE`/`NEXT_HEADING_RE` regexes —
  the block-splicing that wraps the expect block and appends the
  Ready-to-Begin / bottom-nav extras

## Barrel stability

`simpleMarkdownToHtml` and `injectJourneyExtrasIntoJsonGuide` are part of the
public surface (the former is imported by the sibling content-fetcher.test,
the latter by components/docs-panel). content-fetcher.ts re-exports both, so
`export * from './content-fetcher'` keeps the barrel identical — no consumer
changes. The internal `injectJourneyExtrasIntoJsonGuide` call site is rewired
to the import. `sanitizeHtmlUrl` (used only by inlineMarkdown) is dropped from
content-fetcher's security import.

## Tests (+12, gap #5)

New `cover-page.test.ts` pins the block-splicing that was only tested
transitively before: expect-heading wrapping, content-before preserved as its
own block, remainder-after-next-heading preserved, card body pulled from the
following block (and spliced out), first-heading-only, trailing extras block,
invalid-JSON / no-blocks passthrough, and apostrophe-variant heading matching.

## Reviewer gotcha

`EXPECT_HEADING_RE` is byte-identical to the original, including the literal
`’` escape inside the apostrophe character class — verified via hexdump,
not retyped. The pre-existing swallowed-error console noise is left untouched
per the refactor hard gates.
